enclave-tls/dist: support debuginfo packages that can run enclave-tls-server/client successfully in centos 8.2
Because in centos 8.x, rpmbuild has a higher 4.14 version,
and introduces the debugsource sub-packages. So it is not
easy to disable the strip operation, so we customized the
__enclave_tls_install_post and __spec_install_post macros.

The __enclave_tls_install_post macro is executed after
the stripping and we use the sgx_stub_enclave.signed.so which
generated during compilation to replaces the corresponding stripped
one which generated during the install phase.

As a result, we ensure the correct operation of enclave-tls-server/client,
and successfully generate the debuginfo and debugsource subpackages
of enclave-tls

Fixes: #1029
Signed-off-by: Yilin Li <YiLin.Li@linux.alibaba.com>