rats-tls: fix td quote generation failure
If input quote buffer is NULL, the tdx_gen_quote() will return
the required length of quote buffer, so it is wrong to get the
returned NULL quote buffer as result.

In addition, tdx_att_get_quote() uses report data rather than
td report as the first parameter. This means tdx_att_get_report()
currently is not so useful.

Signed-off-by: Jia Zhang <zhang.jia@linux.alibaba.com>